Move project to viper git repos
