+++ src/main/java/vct/main/Main.java
@@ -149,7 +149,7 @@ public class Main
       clops.add(learn.getEnable("Learn unit times for AST nodes."), "learn");
       
       Configuration.add_options(clops);
-      viper.api.util.Configuration.addOptions(clops);
+      viper.api.config.Configuration.addOptions(clops);
 
       String input[]=clops.parse(args);
